+++ Changelog.md
@@ -36,6 +36,7 @@
 * `ParagraphTab` now stored in `Tabs` in `ParagraphProperties`
 * Add `OoxmlSize` support of `one_240th_cm` and use in `Spacing`
 * Add `OoxmlSize` support of `spacing_point` and use in `Spacing`
+* Add `OoxmlSize` support of `one_100th_point` and use in `RunProperties#spacing`
 * Add parsing `PresetColor` to `GradientStop`
 * Add Parsing `NumberingLevel#suffix`
 * Add parsing `CellProperties#no_wrap`